split ggml_quantize_mat_q8_0/K
